Update README.md
Microsoft Office Professional Plus 2021-2024 VL v2411 Build 18227.20152

Office 2024 includes virtually every change that has been added to the suite of programs in Office 365 over the past 3 years, although some features remain exclusive to subscribers. In addition to the main programs Word, Excel, PowerPoint, and Outlook, which are available for both Windows and macOS, the Access and Publisher versions are only available for Windows and are included, as usual, only in premiums. Visio and Project for Windows are also available, although they are not part of the ongoing release of Office.

[Download Microsoft Office Professional Plus 2021-2024 VL v2411 Build 18227.20152](https://developer.team/software/35100-microsoft-office-professional-plus-2021-2024-vl-v2411-build-1822720152.html)